backend/requirements.txt

Django>=4.2,<5.0

mysqlclient>=2.1

djangorestframework>=3.14

djangorestframework-simplejwt>=5.2

django-cors-headers>=3.13

qrcode>=7.3
Pillow>=9.5

#Pour le frontend/Le React
# npm install react react-dom react-router-dom axios \
#   @chakra-ui/react @chakra-ui/icons @emotion/react @emotion/styled \
#   framer-motion react-icons
